+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog for MishkaDeveloperTools 0.1.2
 
+- [x] Add allowed parent domain core key
 - [ ] Add driver for accepting JSON
 - [ ] Add Conditional field structure `macro` (**Multiple states of a field**)
 - [ ] Add driver for making JSON outputs